Moved generic OpState cluster objects into one header file.
