+++ src/app/clusters/operational-state-server/operational-state-delegate.h
@@ -18,6 +18,7 @@
 
 #pragma once
 
+#include "operational-state-cluster-objects.h"
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app/util/af-enums.h>
 #include <lib/support/CommonIterator.h>
@@ -26,166 +27,6 @@ namespace chip {
 namespace app {
 namespace Clusters {
 namespace OperationalState {
-
-constexpr size_t kOperationalStateLabelMaxSize   = 64u;
-constexpr size_t kOperationalErrorLabelMaxSize   = 64u;
-constexpr size_t kOperationalErrorDetailsMaxSize = 64u;
-constexpr size_t kOperationalPhaseNameMaxSize    = 64u;
-
-/**
- * A class which represents the operational state of an Operational State cluster derivation instance.
- */
-struct GenericOperationalState : public app::Clusters::detail::Structs::OperationalStateStruct::Type
-{
-    GenericOperationalState(uint8_t state = to_underlying(OperationalStateEnum::kStopped), Optional<CharSpan> label = NullOptional)
-    {
-        Set(state, label);
-    }
-
-    GenericOperationalState(const GenericOperationalState & op) { *this = op; }
-
-    GenericOperationalState & operator=(const GenericOperationalState & op)
-    {
-        Set(op.operationalStateID, op.operationalStateLabel);
-        return *this;
-    }
-
-    void Set(uint8_t state, Optional<CharSpan> label = NullOptional)
-    {
-        operationalStateID = state;
-        if (label.HasValue())
-        {
-            memset(mOperationalStateLabelBuffer, 0, sizeof(mOperationalStateLabelBuffer));
-            if (label.Value().size() > sizeof(mOperationalStateLabelBuffer))
-            {
-                memcpy(mOperationalStateLabelBuffer, label.Value().data(), sizeof(mOperationalStateLabelBuffer));
-                operationalStateLabel.SetValue(CharSpan(mOperationalStateLabelBuffer, sizeof(mOperationalStateLabelBuffer)));
-            }
-            else
-            {
-                memcpy(mOperationalStateLabelBuffer, label.Value().data(), label.Value().size());
-                operationalStateLabel.SetValue(CharSpan(mOperationalStateLabelBuffer, label.Value().size()));
-            }
-        }
-        else
-        {
-            operationalStateLabel = NullOptional;
-        }
-    }
-
-private:
-    char mOperationalStateLabelBuffer[kOperationalStateLabelMaxSize];
-};
-
-/**
- * A class which represents the operational error of an Operational State cluster derivation instance.
- */
-struct GenericOperationalError : public app::Clusters::detail::Structs::ErrorStateStruct::Type
-{
-    GenericOperationalError(uint8_t state, Optional<chip::CharSpan> label = NullOptional,
-                            Optional<chip::CharSpan> details = NullOptional)
-    {
-        Set(state, label, details);
-    }
-
-    GenericOperationalError(const GenericOperationalError & error) { *this = error; }
-
-    GenericOperationalError & operator=(const GenericOperationalError & error)
-    {
-        Set(error.errorStateID, error.errorStateLabel, error.errorStateDetails);
-        return *this;
-    }
-
-    void Set(uint8_t state, Optional<chip::CharSpan> label = NullOptional, Optional<chip::CharSpan> details = NullOptional)
-    {
-        errorStateID = state;
-        if (label.HasValue())
-        {
-            memset(mErrorStateLabelBuffer, 0, sizeof(mErrorStateLabelBuffer));
-            if (label.Value().size() > sizeof(mErrorStateLabelBuffer))
-            {
-                memcpy(mErrorStateLabelBuffer, label.Value().data(), sizeof(mErrorStateLabelBuffer));
-                errorStateLabel.SetValue(CharSpan(mErrorStateLabelBuffer, sizeof(mErrorStateLabelBuffer)));
-            }
-            else
-            {
-                memcpy(mErrorStateLabelBuffer, label.Value().data(), label.Value().size());
-                errorStateLabel.SetValue(CharSpan(mErrorStateLabelBuffer, label.Value().size()));
-            }
-        }
-        else
-        {
-            errorStateLabel = NullOptional;
-        }
-
-        if (details.HasValue())
-        {
-            memset(mErrorStateDetailsBuffer, 0, sizeof(mErrorStateDetailsBuffer));
-            if (details.Value().size() > sizeof(mErrorStateDetailsBuffer))
-            {
-                memcpy(mErrorStateDetailsBuffer, details.Value().data(), sizeof(mErrorStateDetailsBuffer));
-                errorStateDetails.SetValue(CharSpan(mErrorStateDetailsBuffer, sizeof(mErrorStateDetailsBuffer)));
-            }
-            else
-            {
-                memcpy(mErrorStateDetailsBuffer, details.Value().data(), details.Value().size());
-                errorStateDetails.SetValue(CharSpan(mErrorStateDetailsBuffer, details.Value().size()));
-            }
-        }
-        else
-        {
-            errorStateDetails = NullOptional;
-        }
-    }
-
-private:
-    char mErrorStateLabelBuffer[kOperationalErrorLabelMaxSize];
-    char mErrorStateDetailsBuffer[kOperationalErrorDetailsMaxSize];
-};
-
-/**
- * A class which represents the operational phase of an Operational State cluster derivation instance.
- */
-struct GenericOperationalPhase
-{
-    GenericOperationalPhase(app::DataModel::Nullable<CharSpan> name) { Set(name); }
-
-    GenericOperationalPhase(const GenericOperationalPhase & ph) { *this = ph; }
-
-    GenericOperationalPhase & operator=(const GenericOperationalPhase & ph)
-    {
-        Set(ph.mPhaseName);
-        return *this;
-    }
-
-    bool IsMissing() const { return mPhaseName.IsNull(); }
-    app::DataModel::Nullable<CharSpan> mPhaseName;
-
-private:
-    void Set(app::DataModel::Nullable<CharSpan> name)
-    {
-        if (name.IsNull())
-        {
-            mPhaseName.SetNull();
-        }
-        else
-        {
-            memset(mPhaseNameBuffer, 0, sizeof(mPhaseNameBuffer));
-            if (name.Value().size() > sizeof(mPhaseNameBuffer))
-            {
-                memcpy(mPhaseNameBuffer, name.Value().data(), sizeof(mPhaseNameBuffer));
-                mPhaseName = app::DataModel::Nullable<CharSpan>(CharSpan(mPhaseNameBuffer, sizeof(mPhaseNameBuffer)));
-            }
-            else
-            {
-                memcpy(mPhaseNameBuffer, name.Value().data(), name.Value().size());
-                mPhaseName = app::DataModel::Nullable<CharSpan>(CharSpan(mPhaseNameBuffer, name.Value().size()));
-            }
-        }
-    }
-
-    char mPhaseNameBuffer[kOperationalPhaseNameMaxSize];
-};
 
 /**
  * A delegate to handle application logic of the Operational State aliased Cluster.
